Add files via upload
The VAG Case Study Project focuses on analyzing and developing a solution for a specific business challenge. The goal is to design a streamlined and interactive software/web application that demonstrates modern development practices, with an emphasis on usability, efficiency, and innovation.

This project incorporates key programming techniques such as JavaScript, HTML5, CSS3, and ECMAScript 6 (ES6) features. It aims to showcase proficiency in building dynamic front-end interfaces and handling user interactions seamlessly.